% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{active}
\alias{active}
\title{Get the active component of a tidymatrix}
\usage{
active(.data)
}
\arguments{
\item{.data}{A tidymatrix object}
}
\value{
A character string indicating the active component
}
\description{
Get the active component of a tidymatrix
}
