% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_analysis}
\alias{get_analysis}
\title{Get stored analysis object}
\usage{
get_analysis(x, name)
}
\arguments{
\item{x}{A tidymatrix object}

\item{name}{Name of the analysis to retrieve}
}
\value{
The stored analysis object (class depends on analysis type)
}
\description{
Retrieve a full analysis object (prcomp, hclust, etc.) that was stored
using a compute_* function.
}
\examples{
\dontrun{
tm <- tidymatrix(matrix(rnorm(100), 10, 10))
tm <- tm |>
  activate(columns) |>
  compute_prcomp(name = "pca")

# Get the full prcomp object
pca_obj <- get_analysis(tm, "pca")
summary(pca_obj)
plot(pca_obj$sdev)  # Scree plot
}
}
