% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{get_matrix_fn_suggestions}
\alias{get_matrix_fn_suggestions}
\title{Get suggestions for matrix aggregation functions}
\usage{
get_matrix_fn_suggestions(type)
}
\description{
Get suggestions for matrix aggregation functions
}
\keyword{internal}
