% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{group_vars.grouped_tidymatrix}
\alias{group_vars.grouped_tidymatrix}
\title{Get grouping variable names}
\usage{
\method{group_vars}{grouped_tidymatrix}(x)
}
\arguments{
\item{x}{A grouped_tidymatrix object}
}
\value{
A character vector of grouping variable names
}
\description{
Get grouping variable names
}
