% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{groups.grouped_tidymatrix}
\alias{groups.grouped_tidymatrix}
\title{Get grouping variables}
\usage{
\method{groups}{grouped_tidymatrix}(x)
}
\arguments{
\item{x}{A grouped_tidymatrix object}
}
\value{
A list of grouping variable names
}
\description{
Get grouping variables
}
