% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{is_grouped_tidymatrix}
\alias{is_grouped_tidymatrix}
\title{Check if object is a grouped tidymatrix}
\usage{
is_grouped_tidymatrix(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE if the object is a grouped_tidymatrix, FALSE otherwise
}
\description{
Check if object is a grouped tidymatrix
}
