% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymatrix.R
\name{is_tidymatrix}
\alias{is_tidymatrix}
\title{Check if an object is a tidymatrix}
\usage{
is_tidymatrix(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE if the object is a tidymatrix, FALSE otherwise
}
\description{
Check if an object is a tidymatrix
}
