% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{list_analyses}
\alias{list_analyses}
\title{List stored analyses}
\usage{
list_analyses(x)
}
\arguments{
\item{x}{A tidymatrix object}
}
\value{
Character vector of analysis names
}
\description{
Get the names of all analyses stored in a tidymatrix object.
}
\examples{
\dontrun{
tm <- tidymatrix(matrix(rnorm(100), 10, 10))
tm <- tm |>
  activate(columns) |>
  compute_prcomp(name = "pca") |>
  compute_hclust(k = 3, name = "clusters")

list_analyses(tm)
# [1] "pca" "clusters"
}
}
