% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{print.grouped_tidymatrix}
\alias{print.grouped_tidymatrix}
\title{Print a grouped tidymatrix}
\usage{
\method{print}{grouped_tidymatrix}(x, ...)
}
\arguments{
\item{x}{A grouped_tidymatrix object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print a grouped tidymatrix
}
