% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymatrix.R
\name{print.tidymatrix}
\alias{print.tidymatrix}
\title{Print a tidymatrix object}
\usage{
\method{print}{tidymatrix}(x, ...)
}
\arguments{
\item{x}{A tidymatrix object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print a tidymatrix object
}
