% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-methods-additional.R
\name{pull.tidymatrix}
\alias{pull.tidymatrix}
\title{Extract a column from metadata as a vector}
\usage{
\method{pull}{tidymatrix}(.data, var = -1, ...)
}
\arguments{
\item{.data}{A tidymatrix object}

\item{var}{Column name to extract (can be unquoted)}

\item{...}{Not used}
}
\value{
A vector containing the column values
}
\description{
Extract a single column from the active metadata component as a vector.
}
