% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-methods-additional.R
\name{relocate.tidymatrix}
\alias{relocate.tidymatrix}
\title{Relocate columns in metadata}
\usage{
\method{relocate}{tidymatrix}(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{A tidymatrix object}

\item{...}{Columns to move}

\item{.before}{Column to move before}

\item{.after}{Column to move after}
}
\value{
A tidymatrix object with reordered metadata columns
}
\description{
Change the order of columns in the active metadata component.
}
