% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-methods-additional.R
\name{slice_head.tidymatrix}
\alias{slice_head.tidymatrix}
\alias{slice_tail.tidymatrix}
\title{Select first or last rows/columns}
\usage{
\method{slice_head}{tidymatrix}(.data, n, prop, ...)

\method{slice_tail}{tidymatrix}(.data, n, prop, ...)
}
\arguments{
\item{.data}{A tidymatrix object}

\item{n}{Number of rows/columns to select}

\item{prop}{Proportion of rows/columns to select}

\item{...}{Not used}
}
\value{
A tidymatrix object with selected data
}
\description{
Select first or last rows/columns
}
