% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-methods-additional.R
\name{slice_sample.tidymatrix}
\alias{slice_sample.tidymatrix}
\title{Select a random sample of rows/columns}
\usage{
\method{slice_sample}{tidymatrix}(.data, n, prop, weight_by = NULL, replace = FALSE, ...)
}
\arguments{
\item{.data}{A tidymatrix object}

\item{n}{Number of rows/columns to select}

\item{prop}{Proportion of rows/columns to select}

\item{weight_by}{Sampling weights (not yet implemented)}

\item{replace}{Sample with replacement}

\item{...}{Not used}
}
\value{
A tidymatrix object with sampled data
}
\description{
Select a random sample of rows/columns
}
