% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{ungroup.grouped_tidymatrix}
\alias{ungroup.grouped_tidymatrix}
\title{Remove grouping from a grouped tidymatrix}
\usage{
\method{ungroup}{grouped_tidymatrix}(x, ...)
}
\arguments{
\item{x}{A grouped_tidymatrix object}

\item{...}{Not used}
}
\value{
A tidymatrix object (ungrouped)
}
\description{
Remove grouping from a grouped tidymatrix
}
