% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymatrix.R
\name{validate_tidymatrix}
\alias{validate_tidymatrix}
\title{Validate a tidymatrix object}
\usage{
validate_tidymatrix(tm)
}
\arguments{
\item{tm}{A tidymatrix object}
}
\value{
The tidymatrix object (invisibly) if valid, otherwise throws an error
}
\description{
Validate a tidymatrix object
}
\keyword{internal}
